move 'addIndex' tests to separate file
