+++ types/ramda/ramda-tests.ts
@@ -637,16 +637,8 @@ R.times(i, 5);
         return n % 2 === 0;
     }
 
-    const filterIndexed = R.addIndex(R.filter);
-
     R.filter(isEven, [1, 2, 3, 4]); // => [2, 4]
     R.filter(isEven, { a: 0, b: 1 }); // => { a: 0 }
-
-    function lastTwo(val: number, idx: number, list: number[]) {
-        return list.length - idx <= 2;
-    }
-
-    filterIndexed(lastTwo, [8, 6, 7, 5, 3, 0, 9]); // => [0, 9]
 
     function isOdd(n: number) {
         return n % 2 === 1;
@@ -841,18 +833,6 @@ R.times(i, 5);
 };
 
 () => {
-    function lastTwo(val: number, idx: number, list: number[]) {
-        return list.length - idx <= 2;
-    }
-
-    const filterIndexed = R.addIndex(R.filter);
-
-    filterIndexed(lastTwo, [8, 6, 7, 5, 3, 0, 9]); // => [0, 9]
-    const lastTwoFn = filterIndexed(lastTwo);
-    lastTwoFn([8, 6, 7, 5, 3, 0, 9]);
-};
-
-() => {
     const xs = [{a: 1}, {a: 2}, {a: 3}];
     R.find(R.propEq("a", 2))(xs); // => {a: 2}
     R.find(R.propEq("a", 4))(xs); // => undefined
@@ -953,14 +933,6 @@ interface Obj {
     // -> 6
     // -> 7
     // -> 8
-};
-
-() => {
-    function plusFive(num: number, idx: number, list: number[]) {
-        list[idx] = num + 5;
-    }
-
-    R.addIndex(R.forEach)(plusFive)([1, 2, 3]); // => [6, 7, 8]
 };
 
 () => {
@@ -1181,18 +1153,6 @@ interface Obj {
 };
 
 () => {
-    function squareEnds(elt: number, idx: number, list: number[]) {
-        if (idx === 0 || idx === list.length - 1) {
-            return elt * elt;
-        }
-        return elt;
-    }
-
-    R.addIndex(R.map)(squareEnds, [8, 5, 3, 0, 9]); // => [64, 5, 3, 0, 81]
-    R.addIndex(R.map)(squareEnds)([8, 5, 3, 0, 9]); // => [64, 5, 3, 0, 81]
-};
-
-() => {
     const sampleList = ['a', 'b', 'c', 'd', 'e', 'f'];
 
     R.move<string>(0, 2, sampleList); // => ['b', 'c', 'a', 'd', 'e', 'f']
@@ -1278,20 +1238,6 @@ interface Student {
     // }
 };
 
-() => {
-    const reduceIndexed = R.addIndex(R.reduce);
-    const letters       = ["a", "b", "c"];
-
-    function objectify(accObject: { [elem: string]: number }, elem: string, idx: number, list: string[]) {
-        accObject[elem] = idx;
-        return accObject;
-    }
-
-    reduceIndexed(objectify, {}, letters); // => { 'a': 0, 'b': 1, 'c': 2 }
-    reduceIndexed(objectify)({}, letters); // => { 'a': 0, 'b': 1, 'c': 2 }
-    reduceIndexed(objectify, {})(letters); // => { 'a': 0, 'b': 1, 'c': 2 }
-};
-
 interface KeyValuePair<K, V> extends Array<K | V> {
     0: K;
     1: V;
@@ -1341,16 +1287,6 @@ type Pair = KeyValuePair<string, number>;
 
     R.reject(isOdd, [1, 2, 3, 4]); // => [2, 4]
     R.reject(isOdd)([1, 2, 3, 4]); // => [2, 4]
-};
-
-() => {
-    function lastTwo(val: number, idx: number, list: number[]) {
-        return list.length - idx <= 2;
-    }
-
-    const rejectIndexed = R.addIndex(R.reject);
-    rejectIndexed(lastTwo, [8, 6, 7, 5, 3, 0, 9]); // => [8, 6, 7, 5, 3]
-    rejectIndexed(lastTwo)([8, 6, 7, 5, 3, 0, 9]); // => [8, 6, 7, 5, 3]
 };
 
 () => {
@@ -2206,20 +2142,6 @@ class Rectangle {
 
 () => {
     const a: number[] = R.without([1, 2], [1, 2, 1, 3, 4]); // => [3, 4]
-};
-
-() => {
-    const mapIndexed = R.addIndex(R.map);
-    mapIndexed((val: string, idx: number) => `${idx}-${val}`)(["f", "o", "o", "b", "a", "r"]);
-    // => ['0-f', '1-o', '2-o', '3-b', '4-a', '5-r']
-    mapIndexed((rectangle: Rectangle, idx: number): number => rectangle.area() * idx, [new Rectangle(1, 2), new Rectangle(4, 7)]);
-    // => [2, 56]
-};
-
-() => {
-    const reduceIndexed = R.addIndex(R.reduce);
-    reduceIndexed((acc: string, val: string, idx: number) => `${acc},${idx}-${val}`, "", ["f", "o", "o", "b", "a", "r"]);
-    // => ['0-f,1-o,2-o,3-b,4-a,5-r']
 };
 
 () => {
